Split into multiple module
Refactored to styled components.
